pyproject.toml
[project]
name = "bricoia"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django==6.0.4",
    "django-htmx>=1.17",
    "google-genai>=1.0",
    "gunicorn>=22",
    "psycopg[binary]>=3.2",
    "python-decouple>=3.8",
    "whitenoise>=6.7",
]

[dependency-groups]
dev = [
    "coverage>=7.6",
    "django-debug-toolbar>=4.4",
    "pytest-django>=4.9",
]
